requirements.txt
absl-py==1.0.0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==22.1.0
blis==0.7.5
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
findspark==2.0.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.8.0
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
absl-py==1.0.0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==22.1.0
blis==0.7.5
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
findspark==2.0.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.8.0
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
srsly==2.4.2
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
absl-py==1.0.0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==22.1.0
blis==0.7.5
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
findspark==2.0.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.8.0
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
srsly==2.4.2
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
statsmodels
seaborn
kafka-python
transformers